Split character builder workspace helpers
