+++ frontend/src/components/tools/CharacterBuilderWorkspace.tsx
@@ -26,8 +26,6 @@ import { Card } from '@/components/ui/Card';
 import { Input, Textarea } from '@/components/ui/Input';
 import { SelectMenu } from '@/components/ui/SelectMenu';
 import { authFetch } from '@/lib/authFetch';
-import { dispatchAnalyticsEvent } from '@/lib/analytics-client';
-import { prepareImageFileForUpload } from '@/lib/client-image-upload';
 import { suggestDownloadFilename, triggerAppDownload } from '@/lib/download';
 import { useI18n } from '@/lib/i18n/I18nProvider';
 import {
@@ -35,9 +33,7 @@ import {
   AGE_RANGE_OPTIONS,
   AUTO_TRAIT_KEYS,
   BODY_BUILD_OPTIONS,
-  CHARACTER_BUILDER_MAX_REFERENCE_IMAGES,
   CHARACTER_BUILDER_STORAGE_KEY,
-  CHARACTER_BUILDER_STORAGE_VERSION,
   CHARACTER_CONSISTENCY_OPTIONS,
   CHARACTER_OUTPUT_OPTIONS,
   CHARACTER_QUALITY_OPTIONS,
@@ -52,7 +48,6 @@ import {
   HAIRSTYLE_OPTIONS,
   getAvailableCharacterFormatOptions,
   getCharacterFormatMultiplier,
-  getQualityEngineId,
   normalizeCharacterFormatMode,
   normalizeTraitsForSourceMode,
   OUTFIT_STYLE_OPTIONS,
@@ -64,7 +59,6 @@ import { useRequireAuth } from '@/hooks/useRequireAuth';
 import { FEATURES } from '@/content/feature-flags';
 import type {
   CharacterBuilderAction,
-  CharacterBuilderFormatMode,
   CharacterBuilderResult,
   CharacterBuilderRun,
   CharacterBuilderSettingsSnapshot,
@@ -73,991 +67,56 @@ import type {
   CharacterBuilderTraits,
   CharacterBuilderReferenceImage,
 } from '@/types/character-builder';
-
-type UploadedAsset = {
-  url: string;
-  width?: number | null;
-  height?: number | null;
-  name?: string | null;
-};
-
-type CharacterLibraryAsset = {
-  id: string;
-  url: string;
-  mime?: string | null;
-  width?: number | null;
-  height?: number | null;
-  size?: number | null;
-  source?: string | null;
-  createdAt?: string;
-};
-
-type CharacterLibraryAssetsResponse = {
-  ok?: boolean;
-  assets?: CharacterLibraryAsset[];
-  error?: string;
-};
-
-type PersistedCharacterBuilderState = {
-  version: number;
-  state: CharacterBuilderState;
-};
-
-type ChoiceOption = {
-  id: string;
-  label: string;
-  description?: string;
-  swatch?: string;
-};
-
-type ToggleItem = {
-  id: string;
-  label: string;
-};
-
-type BillingProductResponse = {
-  ok: boolean;
-  product?: {
-    productKey: string;
-    currency: string;
-    unitPriceCents: number;
-  };
-  error?: string;
-};
-
-function emitClientMetric(event: string, payload?: Record<string, unknown>) {
-  dispatchAnalyticsEvent(event, payload);
-}
-
-const DEFAULT_CHARACTER_COPY = {
-  "auto": "Auto",
-  "custom": "Custom",
-  "notSet": "Not set",
-  "open": "Open",
-  "choose": "Choose",
-  "done": "Done",
-  "close": "Close",
-  "add": "Add",
-  "on": "On",
-  "off": "Off",
-  "disabledTitle": "Tools are disabled",
-  "disabledBody": "Enable `FEATURES.workflows.toolsSection` to access this area.",
-  "back": "Back to Tools",
-  "uploadIdentityStart": "Uploading identity reference…",
-  "uploadStyleStart": "Uploading style reference…",
-  "uploadIdentityDone": "Identity reference uploaded.",
-  "uploadStyleDone": "Style reference uploaded.",
-  "uploadFailed": "Upload failed.",
-  "uploadTooLarge": "Image exceeds {maxMB} MB. Compress it or choose a smaller file.",
-  "loadFromJobDone": "Loaded character builder settings from the selected job.",
-  "duplicateDone": "Builder settings copied from that result.",
-  "missingRun": "Character builder run is missing.",
-  "runGenerateOneDone": "Generated a new character reference.",
-  "runGenerateFourDone": "Generated 4 character reference options.",
-  "runFullBodyDone": "Generated a full-body correction from the selected result.",
-  "runLightingDone": "Generated a lighting variation from the selected result.",
-  "runFailed": "Generation failed.",
-  "savedToLibrary": "Saved to Library.",
-  "saveToLibraryFailed": "Failed to save to Library.",
-  "pricingError": "Unable to load tool pricing",
-  "reset": "Reset",
-  "resetDone": "Builder settings reset to defaults.",
-  "authGate": {
-    "title": "Create an account to generate characters",
-    "body": "Guests can browse the builder and inspect public examples, but generation, uploads, and Library actions require an account.",
-    "primary": "Create account",
-    "secondary": "Sign in",
-    "close": "Close"
-  },
-  "top": {
-    "start": "Start",
-    "buildLook": "Build the look",
-    "generate": "Generate",
-    "resultsEyebrow": "Results",
-    "resultsTitle": "Review outputs"
-  },
-  "sourceMode": {
-    "scratchTitle": "Start from scratch",
-    "scratchBody": "Build the character visually.",
-    "imageTitle": "Start from image",
-    "imageBody": "Infer the look from a reference.",
-    "autoTraits": "Auto traits",
-    "scratchNote": "Start from visual traits only. You can switch to image mode any time."
-  },
-  "references": {
-    "identityTitle": "Identity reference",
-    "identityBody": "Upload the face or character you want to anchor.",
-    "styleTitle": "Style inspiration",
-    "styleBody": "Optional. Use a second image for outfit or style direction.",
-    "addInspiration": "Add inspiration image",
-    "addInspirationBody": "Optional outfit or styling reference.",
-    "remove": "Remove"
-  },
-  "library": {
-    "open": "Library",
-    "choose": "Choose from Library",
-    "body": "Use an uploaded or saved image as a reference.",
-    "error": "Failed to load library images",
-    "empty": "No images available in your library yet.",
-    "tabs": {
-      "all": "All",
-      "upload": "Uploads",
-      "generated": "Generated",
-      "character": "Character",
-      "angle": "Angle"
-    }
-  },
-  "sections": {
-    "gender": "Gender presentation",
-    "age": "Age",
-    "customGenderPlaceholder": "Describe the presentation you want to preserve",
-    "hair": "Hair",
-    "hairOpenEditor": "Open the hair editor",
-    "hairEdit": "Edit",
-    "hairClose": "Close",
-    "outfit": "Outfit",
-    "moreOutfits": "More outfits",
-    "customHair": "Custom hair prompt",
-    "customHairPlaceholder": "Describe the hair you want to preserve or generate",
-    "customOutfit": "Custom outfit prompt",
-    "customOutfitPlaceholder": "Describe the clothes or wardrobe you want to see",
-    "accessoriesFeatures": "Accessories & features",
-    "realism": "Realism style",
-    "moreControls": "Advanced controls",
-    "optional": "Optional",
-    "setCount": "{count} set",
-    "skinTone": "Skin tone",
-    "faceCues": "Face cues",
-    "eyeColor": "Eye color",
-    "bodyBuild": "Body build",
-    "consistency": "Consistency",
-    "referenceStrength": "Reference strength",
-    "accessories": "Accessories",
-    "distinctiveFeatures": "Distinctive features",
-    "customDetails": "Custom details prompt",
-    "customDetailsPlaceholder": "Describe accessories, marks, features, or visible identity details",
-    "advancedNotes": "Advanced notes",
-    "advancedNotesPlaceholder": "Optional extra notes",
-    "mustRemainVisible": "Must remain visible",
-    "mustRemainPlaceholder": "Add a detail"
-  },
-  "outputOptions": {
-    "fullBodyRequired": "Full body required",
-    "includeCloseUps": "4 close-ups below",
-    "neutralStudioBackground": "Neutral background",
-    "preserveFacialDetails": "Preserve facial details",
-    "avoid3dRenderLook": "Avoid 3D look"
-  },
-    "generatePanel": {
-      "portraitTitle": "Portrait reference",
-      "portraitBody": "Face-forward portrait anchor",
-    "sheetTitle": "Character sheet",
-    "sheetBody": "4 full-body views + 4 close-ups",
-    "quality": "Quality",
-    "format": "Format",
-    "pricePerImage": "{price} per image",
-    "generateReference": "Generate reference",
-    "generateFour": "Generate 4 options"
-  },
-  "summary": {
-    "dna": "Character DNA",
-    "snapshot": "Live snapshot",
-    "builderBadge": "Builder",
-    "noReferenceImage": "No reference image",
-    "identity": "Identity",
-    "hair": "Hair",
-    "outfit": "Outfit",
-    "style": "Style",
-    "output": "Output",
-    "quality": "Quality",
-    "format": "Format",
-    "autoFromReference": "Auto from reference",
-    "photoreal": "Photoreal"
-  },
-  "hairEditor": {
-    "title": "Hair editor",
-    "body": "Pick only what matters.",
-    "color": "Color",
-    "length": "Length",
-    "style": "Style"
-  },
-  "followUp": {
-    "selectedEyebrow": "Selected result",
-    "selectedTitle": "Current reference",
-    "selectedAlt": "Selected character reference",
-    "baseReference": "Base reference output",
-    "fullBodyRefinement": "Full-body refinement",
-    "lightingRefinement": "Lighting refinement",
-    "importNote": "Import this reference manually into a video model when you are ready.",
-    "pinnedEyebrow": "Pinned base",
-    "pinnedTitle": "Refinement base",
-    "pinnedAlt": "Pinned base reference",
-    "pinnedLabel": "Pinned base",
-    "unpin": "Unpin",
-    "pinEmpty": "Select a result, then pin it if you want a stable base for refinements.",
-    "refineTitle": "Refine",
-    "fixFullBody": "Fix full body",
-    "createLighting": "Create lighting variant",
-    "refineFootnote": "This tool stops at reusable reference image creation."
-  },
-  "resultCard": {
-    "generatedAlt": "Generated character reference",
-    "referenceOutput": "Reference output",
-    "fullBodyFix": "Full-body fix",
-    "lightingVariant": "Lighting variant",
-    "pending": "Rendering",
-    "pendingBody": "This output will appear here as soon as it is ready.",
-    "selected": "Selected",
-    "base": "Base",
-    "unpinBase": "Unpin base",
-    "useAsBase": "Use as base",
-    "download": "Download",
-    "save": "Save",
-    "duplicate": "Duplicate",
-    "select": "Select"
-  },
-  "options": {
-    "gender": {
-      "woman": "Woman",
-      "man": "Man",
-      "androgynous": "Androgynous",
-      "custom": "Custom"
-    },
-    "age": {
-      "teen": "Teen",
-      "young-adult": "Young adult",
-      "adult": "Adult",
-      "mature": "Mature",
-      "senior": "Senior"
-    },
-    "skinTone": {
-      "fair": "Fair",
-      "light": "Light",
-      "medium": "Medium",
-      "olive": "Olive",
-      "deep": "Deep",
-      "rich": "Rich"
-    },
-    "faceCues": {
-      "soft": "Soft",
-      "angular": "Angular",
-      "round": "Round",
-      "defined": "Defined",
-      "strong-jaw": "Strong jaw"
-    },
-    "hairColor": {
-      "black": "Black",
-      "dark-brown": "Dark brown",
-      "brown": "Brown",
-      "blonde": "Blonde",
-      "red": "Red",
-      "gray": "Gray",
-      "fantasy": "Fantasy"
-    },
-    "hairLength": {
-      "short": "Short",
-      "medium": "Medium",
-      "long": "Long",
-      "very-long": "Very long"
-    },
-    "hairstyle": {
-      "straight": "Straight",
-      "wavy-bob": "Wavy bob",
-      "curly": "Curly",
-      "ponytail": "Ponytail",
-      "braids": "Braids",
-      "buzz-cut": "Buzz cut",
-      "afro": "Afro",
-      "tied-back": "Tied back"
-    },
-    "eyeColor": {
-      "brown": "Brown",
-      "hazel": "Hazel",
-      "blue": "Blue",
-      "green": "Green",
-      "gray": "Gray",
-      "amber": "Amber"
-    },
-    "bodyBuild": {
-      "slim": "Slim",
-      "average": "Average",
-      "athletic": "Athletic",
-      "strong": "Strong",
-      "curvy": "Curvy"
-    },
-    "outfit": {
-      "casual": "Casual",
-      "business": "Business",
-      "streetwear": "Streetwear",
-      "formal": "Formal",
-      "luxury": "Luxury",
-      "sci-fi": "Sci-fi",
-      "fantasy": "Fantasy",
-      "tactical": "Tactical"
-    },
-    "realism": {
-      "photoreal": "Photoreal",
-      "cinematic": "Cinematic",
-      "stylized": "Stylized",
-      "animated": "Animated"
-    },
-    "accessories": {
-      "glasses": "Glasses",
-      "sunglasses": "Sunglasses",
-      "earrings": "Earrings",
-      "necklace": "Necklace",
-      "hat": "Hat",
-      "headscarf": "Headscarf"
-    },
-    "distinctive": {
-      "freckles": "Freckles",
-      "scar": "Scar",
-      "tattoo": "Tattoo",
-      "piercing": "Piercing",
-      "beard": "Beard",
-      "beauty-mark": "Beauty mark",
-      "makeup": "Makeup",
-      "wrinkles": "Wrinkles"
-    },
-    "outputMode": {
-      "portrait-reference": {
-        "label": "Portrait reference",
-        "description": "Centered face-forward anchor for later reuse."
-      },
-      "character-sheet": {
-        "label": "Character sheet",
-        "description": "4 full-body angles plus 4 matching close-ups."
-      }
-    },
-    "consistency": {
-      "exploratory": {
-        "label": "Exploratory",
-        "description": "Keep identity recognizable while allowing more variation."
-      },
-      "balanced": {
-        "label": "Balanced",
-        "description": "Aim for stable identity with a little creative flexibility."
-      },
-      "strict": {
-        "label": "Strict",
-        "description": "Bias strongly toward identity preservation across outputs."
-      }
-    },
-    "referenceStrength": {
-      "loose": {
-        "label": "Loose",
-        "description": "Use the reference for general vibe and broad identity cues."
-      },
-      "balanced": {
-        "label": "Balanced",
-        "description": "Preserve the core face and styling while allowing cleanup."
-      },
-      "strong": {
-        "label": "Strong",
-        "description": "Stay close to the uploaded identity and key visible markers."
-      }
-    },
-    "quality": {
-      "draft": {
-        "label": "Standard",
-        "description": "Fast exploratory passes on Nano Banana 2."
-      },
-      "final": {
-        "label": "Pro",
-        "description": "Cleaner export-ready references on Nano Banana Pro."
-      }
-    },
-    "format": {
-      "standard": {
-        "label": "1K",
-        "description": "Base render size."
-      },
-      "2k": {
-        "label": "2K",
-        "description": "Sharper output. 2x price."
-      },
-      "4k": {
-        "label": "4K",
-        "description": "Final-grade output. 3x price."
-      }
-    }
-  }
-} as const;
-
-type CharacterCopy = typeof DEFAULT_CHARACTER_COPY;
-type LoadingRequestKey = 'generate-1' | 'generate-4' | 'full-body-fix' | 'lighting-variant';
-type LoadingRequestCounts = Record<LoadingRequestKey, number>;
-type PendingCharacterRun = {
-  id: string;
-  jobId: string;
-  createdAt: string;
-  action: CharacterBuilderAction;
-  outputMode: CharacterBuilderRun['outputMode'];
-  qualityMode: CharacterBuilderRun['qualityMode'];
-  formatMode: CharacterBuilderFormatMode;
-  generateCount: 1 | 4;
-};
-
-type PersistedPendingCharacterRuns = {
-  version: 1;
-  runs: PendingCharacterRun[];
-};
-
-type CharacterJobPayload = {
-  ok?: boolean;
-  jobId?: string;
-  createdAt?: string;
-  status?: string | null;
-  renderIds?: string[] | null;
-  renderThumbUrls?: string[] | null;
-  settingsSnapshot?: unknown;
-  pricing?: CharacterBuilderRun['pricing'];
-};
-
-type HistoricalCharacterGalleryItem = {
-  id: string;
-  jobId: string;
-  imageUrl: string;
-  thumbUrl: string;
-  engineLabel: string;
-  createdAt: string;
-  prompt: string | null;
-};
-
-const INITIAL_LOADING_REQUEST_COUNTS: LoadingRequestCounts = {
-  'generate-1': 0,
-  'generate-4': 0,
-  'full-body-fix': 0,
-  'lighting-variant': 0,
-};
-
-const DEFAULT_UPLOAD_LIMIT_MB = Number.isFinite(Number(process.env.NEXT_PUBLIC_ASSET_MAX_IMAGE_MB ?? '25'))
-  ? Number(process.env.NEXT_PUBLIC_ASSET_MAX_IMAGE_MB ?? '25')
-  : 25;
-const CHARACTER_BUILDER_PENDING_RUNS_STORAGE_KEY = 'maxvideoai.character-builder.pending-runs.v1';
-
-function isAuthRequiredError(error: unknown): boolean {
-  if (!error || typeof error !== 'object') return false;
-  const record = error as { status?: number; code?: string };
-  return record.status === 401 || record.code === 'auth_required' || record.code === 'UNAUTHORIZED';
-}
-
-function formatTemplate(template: string, values: Record<string, string | number>): string {
-  return Object.entries(values).reduce((result, [key, value]) => {
-    return result.replaceAll(`{${key}}`, String(value));
-  }, template);
-}
-
-function getLoadingRequestKey(action: CharacterBuilderAction, generateCount?: 1 | 4): LoadingRequestKey {
-  if (action === 'generate') {
-    return generateCount === 4 ? 'generate-4' : 'generate-1';
-  }
-  return action;
-}
-
-function incrementLoadingRequestCount(
-  counts: LoadingRequestCounts,
-  key: LoadingRequestKey
-): LoadingRequestCounts {
-  return {
-    ...counts,
-    [key]: (counts[key] ?? 0) + 1,
-  };
-}
-
-function decrementLoadingRequestCount(
-  counts: LoadingRequestCounts,
-  key: LoadingRequestKey
-): LoadingRequestCounts {
-  return {
-    ...counts,
-    [key]: Math.max(0, (counts[key] ?? 0) - 1),
-  };
-}
-
-function getResultActionLabel(copy: CharacterCopy, action: CharacterBuilderAction): string {
-  if (action === 'full-body-fix') return copy.resultCard.fullBodyFix;
-  if (action === 'lighting-variant') return copy.resultCard.lightingVariant;
-  return copy.resultCard.referenceOutput;
-}
-
-function getFormatDisplayLabel(
-  copy: CharacterCopy,
-  formatMode: CharacterBuilderFormatMode,
-  qualityMode: CharacterBuilderState['qualityMode']
-): string {
-  if (formatMode === '4k') return copy.options.format['4k'].label;
-  if (formatMode === '2k') return copy.options.format['2k'].label;
-  return qualityMode === 'final' ? '2K' : '1K';
-}
-
-
-function getCharacterBillingProductKey(qualityMode: CharacterBuilderState['qualityMode']): string {
-  return qualityMode === 'final' ? 'character-final' : 'character-draft';
-}
-
-function formatUsd(value: number | null | undefined): string {
-  if (typeof value !== 'number' || !Number.isFinite(value)) return 'n/a';
-  return `$${value.toFixed(2)}`;
-}
-
-function getUploadTooLargeMessage(copy: CharacterCopy, maxMB: number): string {
-  return formatTemplate(copy.uploadTooLarge, { maxMB });
-}
-
-async function uploadImage(file: File, copy: CharacterCopy): Promise<UploadedAsset> {
-  const preparedFile = await prepareImageFileForUpload(file, {
-    maxBytes: DEFAULT_UPLOAD_LIMIT_MB * 1024 * 1024,
-  });
-
-  const formData = new FormData();
-  formData.set('file', preparedFile, preparedFile.name);
-
-  const response = await authFetch('/api/uploads/image', {
-    method: 'POST',
-    body: formData,
-  });
-
-  const payload = (await response.json().catch(() => null)) as
-    | {
-        ok?: boolean;
-        error?: string;
-        maxMB?: number;
-        asset?: {
-          url: string;
-          width?: number | null;
-          height?: number | null;
-          name?: string | null;
-        };
-      }
-    | null;
-
-  if (!response.ok || !payload?.ok || !payload.asset?.url) {
-    if (payload?.error === 'FILE_TOO_LARGE' || response.status === 413) {
-      throw new Error(getUploadTooLargeMessage(copy, payload?.maxMB ?? DEFAULT_UPLOAD_LIMIT_MB));
-    }
-    throw new Error(payload?.error ?? `Upload failed (${response.status})`);
-  }
-
-  return {
-    url: payload.asset.url,
-    width: payload.asset.width,
-    height: payload.asset.height,
-    name: payload.asset.name,
-  };
-}
-
-function readPersistedState(): CharacterBuilderState | null {
-  if (typeof window === 'undefined') return null;
-
-  try {
-    const raw = window.localStorage.getItem(CHARACTER_BUILDER_STORAGE_KEY);
-    if (!raw) return null;
-    const parsed = JSON.parse(raw) as PersistedCharacterBuilderState | null;
-    if (!parsed || parsed.version !== CHARACTER_BUILDER_STORAGE_VERSION || !parsed.state) return null;
-    const base = createDefaultCharacterBuilderState(parsed.state.sourceMode === 'reference-image' ? 'reference-image' : 'scratch');
-    return {
-      ...base,
-      ...parsed.state,
-      traits: normalizeHairAndOutfitModes({
-        ...base.traits,
-        ...parsed.state.traits,
-      }),
-      formatMode: normalizeCharacterFormatMode(parsed.state.formatMode, parsed.state.qualityMode ?? base.qualityMode),
-      outputOptions: {
-        ...base.outputOptions,
-        ...parsed.state.outputOptions,
-        fullBodyRequired:
-          (parsed.state.outputMode ?? base.outputMode) === 'character-sheet'
-            ? (parsed.state.outputOptions?.fullBodyRequired ?? base.outputOptions.fullBodyRequired)
-            : false,
-      },
-    };
-  } catch {
-    return null;
-  }
-}
-
-function readPersistedPendingRuns(): PendingCharacterRun[] {
-  if (typeof window === 'undefined') return [];
-
-  try {
-    const raw = window.localStorage.getItem(CHARACTER_BUILDER_PENDING_RUNS_STORAGE_KEY);
-    if (!raw) return [];
-    const parsed = JSON.parse(raw) as PersistedPendingCharacterRuns | null;
-    if (!parsed || parsed.version !== 1 || !Array.isArray(parsed.runs)) return [];
-    return parsed.runs.filter(
-      (entry): entry is PendingCharacterRun =>
-        Boolean(
-          entry &&
-            typeof entry.id === 'string' &&
-            entry.id.length &&
-            typeof entry.jobId === 'string' &&
-            entry.jobId.length &&
-            typeof entry.createdAt === 'string' &&
-            entry.createdAt.length
-        )
-    );
-  } catch {
-    return [];
-  }
-}
-
-function writePersistedPendingRuns(runs: PendingCharacterRun[]) {
-  if (typeof window === 'undefined') return;
-  if (!runs.length) {
-    window.localStorage.removeItem(CHARACTER_BUILDER_PENDING_RUNS_STORAGE_KEY);
-    return;
-  }
-  const payload: PersistedPendingCharacterRuns = {
-    version: 1,
-    runs,
-  };
-  window.localStorage.setItem(CHARACTER_BUILDER_PENDING_RUNS_STORAGE_KEY, JSON.stringify(payload));
-}
-
-function writePersistedState(state: CharacterBuilderState) {
-  if (typeof window === 'undefined') return;
-  const payload: PersistedCharacterBuilderState = {
-    version: CHARACTER_BUILDER_STORAGE_VERSION,
-    state,
-  };
-  window.localStorage.setItem(CHARACTER_BUILDER_STORAGE_KEY, JSON.stringify(payload));
-}
-
-function normalizeTag(value: string): string {
-  return value.trim().replace(/\s+/g, ' ');
-}
-
-function getFlattenedResults(runs: CharacterBuilderRun[]): CharacterBuilderResult[] {
-  return runs.flatMap((run) => run.results);
-}
-
-function getCharacterEngineLabel(qualityMode: CharacterBuilderRun['qualityMode']): string {
-  return getQualityEngineId(qualityMode) === 'nano-banana-pro' ? 'Nano Banana Pro' : 'Nano Banana 2';
-}
-
-function buildRecoveredRunFromJob(
-  pendingRun: PendingCharacterRun,
-  payload: CharacterJobPayload
-): CharacterBuilderRun | null {
-  const renderIds = Array.isArray(payload.renderIds)
-    ? payload.renderIds.filter((value): value is string => typeof value === 'string' && value.length > 0)
-    : [];
-  if (!renderIds.length) return null;
-
-  const renderThumbUrls = Array.isArray(payload.renderThumbUrls)
-    ? payload.renderThumbUrls.filter((value): value is string => typeof value === 'string' && value.length > 0)
-    : [];
-  const snapshot =
-    payload.settingsSnapshot && typeof payload.settingsSnapshot === 'object'
-      ? (payload.settingsSnapshot as CharacterBuilderSettingsSnapshot)
-      : null;
-  const createdAt = payload.createdAt ?? pendingRun.createdAt;
-  const engineId = snapshot?.engineId ?? getQualityEngineId(pendingRun.qualityMode);
-  const engineLabel = snapshot?.engineLabel ?? getCharacterEngineLabel(pendingRun.qualityMode);
-  const runId = `character_run_${pendingRun.jobId}`;
-  const defaultBuilderState = createDefaultCharacterBuilderState();
-  const fallbackSnapshot: CharacterBuilderSettingsSnapshot = {
-    schemaVersion: 1,
-    surface: 'character-builder',
-    action: pendingRun.action,
-    engineId,
-    engineLabel,
-    inputMode: 't2i',
-    prompt: '',
-    builder: {
-      sourceMode: defaultBuilderState.sourceMode,
-      outputMode: pendingRun.outputMode,
-      consistencyMode: defaultBuilderState.consistencyMode,
-      referenceStrength: defaultBuilderState.referenceStrength,
-      qualityMode: pendingRun.qualityMode,
-      formatMode: pendingRun.formatMode,
-      referenceImages: defaultBuilderState.referenceImages,
-      traits: defaultBuilderState.traits,
-      outputOptions: defaultBuilderState.outputOptions,
-      advancedNotes: defaultBuilderState.advancedNotes,
-      mustRemainVisible: defaultBuilderState.mustRemainVisible,
-    },
-    lineage: {
-      parentResultId: null,
-      parentRunId: null,
-      pinnedReferenceResultId: null,
-    },
-  };
-
-  return {
-    id: runId,
-    jobId: pendingRun.jobId,
-    action: snapshot?.action ?? pendingRun.action,
-    outputMode: snapshot?.builder.outputMode ?? pendingRun.outputMode,
-    qualityMode: snapshot?.builder.qualityMode ?? pendingRun.qualityMode,
-    formatMode: normalizeCharacterFormatMode(snapshot?.builder.formatMode ?? pendingRun.formatMode, snapshot?.builder.qualityMode ?? pendingRun.qualityMode),
-    engineId,
-    engineLabel,
-    createdAt,
-    parentResultId: snapshot?.lineage?.parentResultId ?? null,
-    results: renderIds.map((url, index) => ({
-      id: `${pendingRun.jobId}:result:${index + 1}`,
-      runId,
-      jobId: pendingRun.jobId,
-      url,
-      thumbUrl: renderThumbUrls[index] ?? url,
-      engineId,
-      engineLabel,
-      action: snapshot?.action ?? pendingRun.action,
-      outputMode: snapshot?.builder.outputMode ?? pendingRun.outputMode,
-      qualityMode: snapshot?.builder.qualityMode ?? pendingRun.qualityMode,
-      createdAt,
-      parentResultId: snapshot?.lineage?.parentResultId ?? null,
-    })),
-    settingsSnapshot: snapshot ?? fallbackSnapshot,
-    pricing: payload.pricing,
-    requestId: pendingRun.jobId,
-    providerJobId: null,
-  };
-}
-
-function buildReferenceImage(
-  role: CharacterBuilderReferenceImage['role'],
-  asset: UploadedAsset
-): CharacterBuilderReferenceImage {
-  return {
-    id: `${role}_${Date.now()}`,
-    role,
-    url: asset.url,
-    width: asset.width ?? null,
-    height: asset.height ?? null,
-    name: asset.name ?? null,
-  };
-}
-
-function getRefByRole(
-  referenceImages: CharacterBuilderReferenceImage[],
-  role: CharacterBuilderReferenceImage['role']
-) {
-  return referenceImages.find((image) => image.role === role) ?? null;
-}
-
-function updateReferenceImage(
-  referenceImages: CharacterBuilderReferenceImage[],
-  nextImage: CharacterBuilderReferenceImage
-) {
-  const filtered = referenceImages.filter((image) => image.role !== nextImage.role);
-  return [...filtered, nextImage].slice(0, CHARACTER_BUILDER_MAX_REFERENCE_IMAGES);
-}
-
-function removeReferenceImage(
-  referenceImages: CharacterBuilderReferenceImage[],
-  role: CharacterBuilderReferenceImage['role']
-) {
-  return referenceImages.filter((image) => image.role !== role);
-}
-
-function hasCustomHairSettings(traits: CharacterBuilderTraits): boolean {
-  if (typeof traits.customHairDescription === 'string' && traits.customHairDescription.trim().length > 0) {
-    return true;
-  }
-  return [traits.hairColor.value, traits.hairLength.value, traits.hairstyle.value].some(
-    (value) => typeof value === 'string' && value !== 'auto'
-  );
-}
-
-function hasCustomOutfitSettings(traits: CharacterBuilderTraits): boolean {
-  if (typeof traits.customOutfitDescription === 'string' && traits.customOutfitDescription.trim().length > 0) {
-    return true;
-  }
-  return typeof traits.outfitStyle.value === 'string' && traits.outfitStyle.value !== 'auto';
-}
-
-function normalizeHairAndOutfitModes(traits: CharacterBuilderTraits): CharacterBuilderTraits {
-  return {
-    ...traits,
-    hairEnabled: hasCustomHairSettings(traits),
-    outfitEnabled: hasCustomOutfitSettings(traits),
-  };
-}
-
-function parseCharacterBuilderSnapshot(snapshot: unknown): Partial<CharacterBuilderState> | null {
-  if (!snapshot || typeof snapshot !== 'object') return null;
-  const record = snapshot as {
-    schemaVersion?: unknown;
-    surface?: unknown;
-    builder?: unknown;
-    lineage?: unknown;
-  };
-  if (record.schemaVersion !== 1 || record.surface !== 'character-builder') return null;
-
-  const builder = record.builder as CharacterBuilderSettingsSnapshot['builder'] | undefined;
-  const lineage = record.lineage as CharacterBuilderSettingsSnapshot['lineage'] | undefined;
-  if (!builder) return null;
-  const base = createDefaultCharacterBuilderState(builder.sourceMode);
-
-  return {
-    sourceMode: builder.sourceMode ?? base.sourceMode,
-    referenceImages: Array.isArray(builder.referenceImages) ? builder.referenceImages : [],
-    traits: normalizeHairAndOutfitModes({
-      ...base.traits,
-      ...builder.traits,
-    }),
-    outputMode: builder.outputMode ?? base.outputMode,
-    consistencyMode: builder.consistencyMode ?? base.consistencyMode,
-    referenceStrength: builder.referenceStrength ?? base.referenceStrength,
-    qualityMode: builder.qualityMode ?? base.qualityMode,
-    formatMode: normalizeCharacterFormatMode(builder.formatMode, builder.qualityMode ?? base.qualityMode),
-    outputOptions: {
-      ...base.outputOptions,
-      ...builder.outputOptions,
-      fullBodyRequired:
-        (builder.outputMode ?? base.outputMode) === 'character-sheet'
-          ? (builder.outputOptions?.fullBodyRequired ?? base.outputOptions.fullBodyRequired)
-          : false,
-    },
-    advancedNotes: builder.advancedNotes ?? '',
-    mustRemainVisible: Array.isArray(builder.mustRemainVisible) ? builder.mustRemainVisible : [],
-    selectedResultId: typeof lineage?.parentResultId === 'string' ? lineage.parentResultId : null,
-    pinnedReferenceResultId:
-      typeof lineage?.pinnedReferenceResultId === 'string' ? lineage.pinnedReferenceResultId : null,
-  };
-}
-
-function findChoiceLabel(options: Array<{ id: string; label: string }>, value: string | null | undefined): string | null {
-  if (!value) return null;
-  return options.find((option) => option.id === value)?.label ?? null;
-}
-
-function describeTraitValue(
-  options: Array<{ id: string; label: string }>,
-  value: string | null | undefined,
-  copy: CharacterCopy
-): string {
-  if (value === 'auto') return copy.auto;
-  return findChoiceLabel(options, value) ?? copy.notSet;
-}
-
-function summarizeCustomText(value: string | null | undefined): string {
-  const trimmed = value?.trim() ?? '';
-  if (!trimmed) return '';
-  return trimmed.length > 44 ? `${trimmed.slice(0, 41).trim()}...` : trimmed;
-}
-
-function findChoiceSwatch(
-  options: Array<{ id: string; swatch?: string }>,
-  value: string | null | undefined
-): string | null {
-  if (!value || value === 'auto') return null;
-  return options.find((option) => option.id === value)?.swatch ?? null;
-}
-
-function getHairSummary(
-  traits: CharacterBuilderTraits,
-  options: {
-    hairColor: Array<{ id: string; label: string }>;
-    hairLength: Array<{ id: string; label: string }>;
-    hairstyle: Array<{ id: string; label: string }>;
-  },
-  copy: CharacterCopy
-): string {
-  if (!hasCustomHairSettings(traits)) {
-    return copy.auto;
-  }
-
-  const customSummary = summarizeCustomText(traits.customHairDescription);
-  if (customSummary) {
-    return customSummary;
-  }
-
-  const values = [
-    describeTraitValue(options.hairColor, traits.hairColor.value, copy),
-    describeTraitValue(options.hairLength, traits.hairLength.value, copy),
-    describeTraitValue(options.hairstyle, traits.hairstyle.value, copy),
-  ];
-
-  const filteredValues = values.filter(
-    (value, index, array) => value !== copy.notSet || array.every((entry) => entry === copy.notSet)
-  );
-
-  const meaningfulValues = filteredValues.filter((value) => value !== copy.notSet);
-  if (!meaningfulValues.length) return copy.notSet;
-  return meaningfulValues.join(' / ');
-}
-
-function getOutfitSummary(
-  traits: CharacterBuilderTraits,
-  options: Array<{ id: string; label: string }>,
-  copy: CharacterCopy
-): string {
-  if (!hasCustomOutfitSettings(traits)) {
-    return copy.auto;
-  }
-
-  const customSummary = summarizeCustomText(traits.customOutfitDescription);
-  if (customSummary) {
-    return customSummary;
-  }
-
-  return describeTraitValue(options, traits.outfitStyle.value, copy);
-}
-
-function countConfiguredSecondaryControls(state: CharacterBuilderState, hasIdentityReference: boolean): number {
-  let count = 0;
-
-  const traitValues: Array<string | null> = [
-    state.traits.skinTone.value,
-    state.traits.faceCues.value,
-    state.traits.eyeColor.value,
-    state.traits.bodyBuild.value,
-  ];
-  count += traitValues.filter((value) => value != null && value !== 'auto').length;
-  count += state.traits.accessories.length;
-  count += state.traits.distinctiveFeatures.length;
-  if (state.traits.customDetailsDescription?.trim().length) count += 1;
-  count += state.mustRemainVisible.length;
-  if (state.consistencyMode !== 'balanced') count += 1;
-  if (hasIdentityReference && state.referenceStrength && state.referenceStrength !== 'balanced') count += 1;
-  if (state.advancedNotes.trim().length) count += 1;
-  if (state.outputMode === 'character-sheet' && !state.outputOptions.includeCloseUps) count += 1;
-  if (!state.outputOptions.neutralStudioBackground) count += 1;
-  if (!state.outputOptions.preserveFacialDetails) count += 1;
-  if (!state.outputOptions.avoid3dRenderLook) count += 1;
-
-  return count;
-}
-
-function buildResetCharacterBuilderState(state: CharacterBuilderState): CharacterBuilderState {
-  const defaults = createDefaultCharacterBuilderState(state.sourceMode);
-  return {
-    ...state,
-    traits: defaults.traits,
-    outputMode: defaults.outputMode,
-    consistencyMode: defaults.consistencyMode,
-    referenceStrength: defaults.referenceStrength,
-    qualityMode: defaults.qualityMode,
-    formatMode: defaults.formatMode,
-    outputOptions: defaults.outputOptions,
-    advancedNotes: defaults.advancedNotes,
-    mustRemainVisible: defaults.mustRemainVisible,
-    selectedResultId: null,
-    pinnedReferenceResultId: null,
-  };
-}
-
-function serializeResettableCharacterBuilderState(state: CharacterBuilderState): string {
-  return JSON.stringify({
-    sourceMode: state.sourceMode,
-    traits: state.traits,
-    outputMode: state.outputMode,
-    consistencyMode: state.consistencyMode,
-    referenceStrength: state.referenceStrength,
-    qualityMode: state.qualityMode,
-    formatMode: state.formatMode,
-    outputOptions: state.outputOptions,
-    advancedNotes: state.advancedNotes,
-    mustRemainVisible: state.mustRemainVisible,
-    selectedResultId: state.selectedResultId,
-    pinnedReferenceResultId: state.pinnedReferenceResultId,
-  });
-}
+import { DEFAULT_CHARACTER_COPY, type CharacterCopy } from './character-builder/_lib/character-builder-copy';
+import {
+  buildRecoveredRunFromJob,
+  buildReferenceImage,
+  buildResetCharacterBuilderState,
+  CHARACTER_BUILDER_PENDING_RUNS_STORAGE_KEY,
+  countConfiguredSecondaryControls,
+  decrementLoadingRequestCount,
+  emitClientMetric,
+  findChoiceLabel,
+  findChoiceSwatch,
+  formatUsd,
+  getCharacterBillingProductKey,
+  getFlattenedResults,
+  getFormatDisplayLabel,
+  getHairSummary,
+  getLoadingRequestKey,
+  getOutfitSummary,
+  getRefByRole,
+  getResultActionLabel,
+  hasCustomHairSettings,
+  hasCustomOutfitSettings,
+  incrementLoadingRequestCount,
+  INITIAL_LOADING_REQUEST_COUNTS,
+  isAuthRequiredError,
+  normalizeHairAndOutfitModes,
+  normalizeTag,
+  parseCharacterBuilderSnapshot,
+  readPersistedPendingRuns,
+  readPersistedState,
+  removeReferenceImage,
+  serializeResettableCharacterBuilderState,
+  summarizeCustomText,
+  updateReferenceImage,
+  uploadImage,
+  writePersistedPendingRuns,
+  writePersistedState,
+} from './character-builder/_lib/character-builder-helpers';
+import type {
+  BillingProductResponse,
+  CharacterJobPayload,
+  CharacterLibraryAsset,
+  CharacterLibraryAssetsResponse,
+  ChoiceOption,
+  HistoricalCharacterGalleryItem,
+  LoadingRequestCounts,
+  LoadingRequestKey,
+  PendingCharacterRun,
+  ToggleItem,
+} from './character-builder/_lib/character-builder-types';
 
 const GENDER_CARD_META: Record<string, { glyph: string; background: string; accent: string }> = {
   woman: {
