+++ frontend/src/components/tools/character-builder/_lib/character-builder-copy.ts
@@ -1,0 +1,356 @@
+export const DEFAULT_CHARACTER_COPY = {
+  "auto": "Auto",
+  "custom": "Custom",
+  "notSet": "Not set",
+  "open": "Open",
+  "choose": "Choose",
+  "done": "Done",
+  "close": "Close",
+  "add": "Add",
+  "on": "On",
+  "off": "Off",
+  "disabledTitle": "Tools are disabled",
+  "disabledBody": "Enable `FEATURES.workflows.toolsSection` to access this area.",
+  "back": "Back to Tools",
+  "uploadIdentityStart": "Uploading identity reference…",
+  "uploadStyleStart": "Uploading style reference…",
+  "uploadIdentityDone": "Identity reference uploaded.",
+  "uploadStyleDone": "Style reference uploaded.",
+  "uploadFailed": "Upload failed.",
+  "uploadTooLarge": "Image exceeds {maxMB} MB. Compress it or choose a smaller file.",
+  "loadFromJobDone": "Loaded character builder settings from the selected job.",
+  "duplicateDone": "Builder settings copied from that result.",
+  "missingRun": "Character builder run is missing.",
+  "runGenerateOneDone": "Generated a new character reference.",
+  "runGenerateFourDone": "Generated 4 character reference options.",
+  "runFullBodyDone": "Generated a full-body correction from the selected result.",
+  "runLightingDone": "Generated a lighting variation from the selected result.",
+  "runFailed": "Generation failed.",
+  "savedToLibrary": "Saved to Library.",
+  "saveToLibraryFailed": "Failed to save to Library.",
+  "pricingError": "Unable to load tool pricing",
+  "reset": "Reset",
+  "resetDone": "Builder settings reset to defaults.",
+  "authGate": {
+    "title": "Create an account to generate characters",
+    "body": "Guests can browse the builder and inspect public examples, but generation, uploads, and Library actions require an account.",
+    "primary": "Create account",
+    "secondary": "Sign in",
+    "close": "Close"
+  },
+  "top": {
+    "start": "Start",
+    "buildLook": "Build the look",
+    "generate": "Generate",
+    "resultsEyebrow": "Results",
+    "resultsTitle": "Review outputs"
+  },
+  "sourceMode": {
+    "scratchTitle": "Start from scratch",
+    "scratchBody": "Build the character visually.",
+    "imageTitle": "Start from image",
+    "imageBody": "Infer the look from a reference.",
+    "autoTraits": "Auto traits",
+    "scratchNote": "Start from visual traits only. You can switch to image mode any time."
+  },
+  "references": {
+    "identityTitle": "Identity reference",
+    "identityBody": "Upload the face or character you want to anchor.",
+    "styleTitle": "Style inspiration",
+    "styleBody": "Optional. Use a second image for outfit or style direction.",
+    "addInspiration": "Add inspiration image",
+    "addInspirationBody": "Optional outfit or styling reference.",
+    "remove": "Remove"
+  },
+  "library": {
+    "open": "Library",
+    "choose": "Choose from Library",
+    "body": "Use an uploaded or saved image as a reference.",
+    "error": "Failed to load library images",
+    "empty": "No images available in your library yet.",
+    "tabs": {
+      "all": "All",
+      "upload": "Uploads",
+      "generated": "Generated",
+      "character": "Character",
+      "angle": "Angle"
+    }
+  },
+  "sections": {
+    "gender": "Gender presentation",
+    "age": "Age",
+    "customGenderPlaceholder": "Describe the presentation you want to preserve",
+    "hair": "Hair",
+    "hairOpenEditor": "Open the hair editor",
+    "hairEdit": "Edit",
+    "hairClose": "Close",
+    "outfit": "Outfit",
+    "moreOutfits": "More outfits",
+    "customHair": "Custom hair prompt",
+    "customHairPlaceholder": "Describe the hair you want to preserve or generate",
+    "customOutfit": "Custom outfit prompt",
+    "customOutfitPlaceholder": "Describe the clothes or wardrobe you want to see",
+    "accessoriesFeatures": "Accessories & features",
+    "realism": "Realism style",
+    "moreControls": "Advanced controls",
+    "optional": "Optional",
+    "setCount": "{count} set",
+    "skinTone": "Skin tone",
+    "faceCues": "Face cues",
+    "eyeColor": "Eye color",
+    "bodyBuild": "Body build",
+    "consistency": "Consistency",
+    "referenceStrength": "Reference strength",
+    "accessories": "Accessories",
+    "distinctiveFeatures": "Distinctive features",
+    "customDetails": "Custom details prompt",
+    "customDetailsPlaceholder": "Describe accessories, marks, features, or visible identity details",
+    "advancedNotes": "Advanced notes",
+    "advancedNotesPlaceholder": "Optional extra notes",
+    "mustRemainVisible": "Must remain visible",
+    "mustRemainPlaceholder": "Add a detail"
+  },
+  "outputOptions": {
+    "fullBodyRequired": "Full body required",
+    "includeCloseUps": "4 close-ups below",
+    "neutralStudioBackground": "Neutral background",
+    "preserveFacialDetails": "Preserve facial details",
+    "avoid3dRenderLook": "Avoid 3D look"
+  },
+    "generatePanel": {
+      "portraitTitle": "Portrait reference",
+      "portraitBody": "Face-forward portrait anchor",
+    "sheetTitle": "Character sheet",
+    "sheetBody": "4 full-body views + 4 close-ups",
+    "quality": "Quality",
+    "format": "Format",
+    "pricePerImage": "{price} per image",
+    "generateReference": "Generate reference",
+    "generateFour": "Generate 4 options"
+  },
+  "summary": {
+    "dna": "Character DNA",
+    "snapshot": "Live snapshot",
+    "builderBadge": "Builder",
+    "noReferenceImage": "No reference image",
+    "identity": "Identity",
+    "hair": "Hair",
+    "outfit": "Outfit",
+    "style": "Style",
+    "output": "Output",
+    "quality": "Quality",
+    "format": "Format",
+    "autoFromReference": "Auto from reference",
+    "photoreal": "Photoreal"
+  },
+  "hairEditor": {
+    "title": "Hair editor",
+    "body": "Pick only what matters.",
+    "color": "Color",
+    "length": "Length",
+    "style": "Style"
+  },
+  "followUp": {
+    "selectedEyebrow": "Selected result",
+    "selectedTitle": "Current reference",
+    "selectedAlt": "Selected character reference",
+    "baseReference": "Base reference output",
+    "fullBodyRefinement": "Full-body refinement",
+    "lightingRefinement": "Lighting refinement",
+    "importNote": "Import this reference manually into a video model when you are ready.",
+    "pinnedEyebrow": "Pinned base",
+    "pinnedTitle": "Refinement base",
+    "pinnedAlt": "Pinned base reference",
+    "pinnedLabel": "Pinned base",
+    "unpin": "Unpin",
+    "pinEmpty": "Select a result, then pin it if you want a stable base for refinements.",
+    "refineTitle": "Refine",
+    "fixFullBody": "Fix full body",
+    "createLighting": "Create lighting variant",
+    "refineFootnote": "This tool stops at reusable reference image creation."
+  },
+  "resultCard": {
+    "generatedAlt": "Generated character reference",
+    "referenceOutput": "Reference output",
+    "fullBodyFix": "Full-body fix",
+    "lightingVariant": "Lighting variant",
+    "pending": "Rendering",
+    "pendingBody": "This output will appear here as soon as it is ready.",
+    "selected": "Selected",
+    "base": "Base",
+    "unpinBase": "Unpin base",
+    "useAsBase": "Use as base",
+    "download": "Download",
+    "save": "Save",
+    "duplicate": "Duplicate",
+    "select": "Select"
+  },
+  "options": {
+    "gender": {
+      "woman": "Woman",
+      "man": "Man",
+      "androgynous": "Androgynous",
+      "custom": "Custom"
+    },
+    "age": {
+      "teen": "Teen",
+      "young-adult": "Young adult",
+      "adult": "Adult",
+      "mature": "Mature",
+      "senior": "Senior"
+    },
+    "skinTone": {
+      "fair": "Fair",
+      "light": "Light",
+      "medium": "Medium",
+      "olive": "Olive",
+      "deep": "Deep",
+      "rich": "Rich"
+    },
+    "faceCues": {
+      "soft": "Soft",
+      "angular": "Angular",
+      "round": "Round",
+      "defined": "Defined",
+      "strong-jaw": "Strong jaw"
+    },
+    "hairColor": {
+      "black": "Black",
+      "dark-brown": "Dark brown",
+      "brown": "Brown",
+      "blonde": "Blonde",
+      "red": "Red",
+      "gray": "Gray",
+      "fantasy": "Fantasy"
+    },
+    "hairLength": {
+      "short": "Short",
+      "medium": "Medium",
+      "long": "Long",
+      "very-long": "Very long"
+    },
+    "hairstyle": {
+      "straight": "Straight",
+      "wavy-bob": "Wavy bob",
+      "curly": "Curly",
+      "ponytail": "Ponytail",
+      "braids": "Braids",
+      "buzz-cut": "Buzz cut",
+      "afro": "Afro",
+      "tied-back": "Tied back"
+    },
+    "eyeColor": {
+      "brown": "Brown",
+      "hazel": "Hazel",
+      "blue": "Blue",
+      "green": "Green",
+      "gray": "Gray",
+      "amber": "Amber"
+    },
+    "bodyBuild": {
+      "slim": "Slim",
+      "average": "Average",
+      "athletic": "Athletic",
+      "strong": "Strong",
+      "curvy": "Curvy"
+    },
+    "outfit": {
+      "casual": "Casual",
+      "business": "Business",
+      "streetwear": "Streetwear",
+      "formal": "Formal",
+      "luxury": "Luxury",
+      "sci-fi": "Sci-fi",
+      "fantasy": "Fantasy",
+      "tactical": "Tactical"
+    },
+    "realism": {
+      "photoreal": "Photoreal",
+      "cinematic": "Cinematic",
+      "stylized": "Stylized",
+      "animated": "Animated"
+    },
+    "accessories": {
+      "glasses": "Glasses",
+      "sunglasses": "Sunglasses",
+      "earrings": "Earrings",
+      "necklace": "Necklace",
+      "hat": "Hat",
+      "headscarf": "Headscarf"
+    },
+    "distinctive": {
+      "freckles": "Freckles",
+      "scar": "Scar",
+      "tattoo": "Tattoo",
+      "piercing": "Piercing",
+      "beard": "Beard",
+      "beauty-mark": "Beauty mark",
+      "makeup": "Makeup",
+      "wrinkles": "Wrinkles"
+    },
+    "outputMode": {
+      "portrait-reference": {
+        "label": "Portrait reference",
+        "description": "Centered face-forward anchor for later reuse."
+      },
+      "character-sheet": {
+        "label": "Character sheet",
+        "description": "4 full-body angles plus 4 matching close-ups."
+      }
+    },
+    "consistency": {
+      "exploratory": {
+        "label": "Exploratory",
+        "description": "Keep identity recognizable while allowing more variation."
+      },
+      "balanced": {
+        "label": "Balanced",
+        "description": "Aim for stable identity with a little creative flexibility."
+      },
+      "strict": {
+        "label": "Strict",
+        "description": "Bias strongly toward identity preservation across outputs."
+      }
+    },
+    "referenceStrength": {
+      "loose": {
+        "label": "Loose",
+        "description": "Use the reference for general vibe and broad identity cues."
+      },
+      "balanced": {
+        "label": "Balanced",
+        "description": "Preserve the core face and styling while allowing cleanup."
+      },
+      "strong": {
+        "label": "Strong",
+        "description": "Stay close to the uploaded identity and key visible markers."
+      }
+    },
+    "quality": {
+      "draft": {
+        "label": "Standard",
+        "description": "Fast exploratory passes on Nano Banana 2."
+      },
+      "final": {
+        "label": "Pro",
+        "description": "Cleaner export-ready references on Nano Banana Pro."
+      }
+    },
+    "format": {
+      "standard": {
+        "label": "1K",
+        "description": "Base render size."
+      },
+      "2k": {
+        "label": "2K",
+        "description": "Sharper output. 2x price."
+      },
+      "4k": {
+        "label": "4K",
+        "description": "Final-grade output. 3x price."
+      }
+    }
+  }
+} as const;
+
+export type CharacterCopy = typeof DEFAULT_CHARACTER_COPY;
